MDEV-31448 Killing a replica thread awaiting its GCO can hang/crash a parallel replica
When a parallel slave worker thread gets killed being in
wait-for-prior-commit stage it may end up cleaning common with prior in commit
order worker resources.

The situation can be perceived as a killed-error sub-case of
MDEV-30780 and then should be tacked uniformly.
To follow suit these fixes make sure a killed worker never breaks tie,
which ensued some simplification in the source code,
with the prior commit "parent", waits for it commit before attends
to the gco critical section.

Tests demonstrate three type of crash/asserts.