MDEV-31872: Add --old_mode=COMPAT_DISCOURAGED
This mode can be used to disable warnings for functions whose use is
strongly discouraged because of inherent limitations in the functionality
they provide. But the use of which can be needed for backwards compatibility
with legacy applications, and which cannot easily be replaced by the user.

Currently used for weak encryption functions ENCODE()/DECODE(), to help
users who need these functions to access existing data encoded with this
non-standard algorithm.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>