MDEV-31995 Bogus error executing PS for query using CTE with renaming of columns
This patch addresses column naming issues with CTEs in the use of prepared
statements and procedures.

Processing statements of the form
with cte (col1, col2) as (select a as c1, b as c2 from t) ...
causes column t.a to be renamed as c1, then col1.

The timing of this renaming is critical and when executing these statements
in both prepared statement and procedures, can produce unexpected results.

with cte (col1, col2) as (select a as c1, b as c2 from t order by c1) ...

This requires column t1.a to be named a when retrieving data, c1 when the
ordering is processed, and col1 when the data is sent to the user.

Things are complicated further when these are processed at prepared statements.
After statement preparation, t1.a is named col1 and needs to be reset to c1
before the statement can be executed at all, leading to the error noted above.

Optimization can lead to second execution errors by pushing conditions into
all the selects in a union, so all the selects will need reninitializing.

In addition to these issues, new column names were allocated
(from Item::set_name() calling thd->alloc()) at each renaming causing memory
leaks. Both the orginal column names and their new names all need to be
allocated in statement memory during preparation.