MDEV-4991: Implement async code path for GTID indexes
When a GTID is written while holding LOCK_log, and a new GTID index record
should be output, defer the index update to the binlog background thread,
minimizing the impact on the binlog performance.

The xid_count_per_binlog::notify_count is no longer needed, as a single
xid_count_per_binlog object can now be queued multiple times in a single
list (the next pointer is now outside of the xid_count_per_binlog).

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>